style(RegistryV2): re-design the export page to have the export form in a modal, plus some minor improvements
