+++ front/src/dashboard/registry/MyExports.module.scss
@@ -1,5 +1,5 @@
 .myExportsForm {
-  max-width: 40rem;
+  max-width: 30rem;
 }
 
 .myRegistryExportsContainer {
