+++ front/src/dashboard/registry/MyImports.tsx
@@ -53,13 +53,13 @@ export function MyImports() {
     const link = await getDownloadLink({
       variables: { importId, target: RegistryDownloadTarget.ErrorFile }
     });
-    await downloadFromSignedUrl(link.data?.registryDownloadSignedUrl.signedUrl);
+    downloadFromSignedUrl(link.data?.registryDownloadSignedUrl.signedUrl);
   }
   async function downloadImportFile(importId: string) {
     const link = await getDownloadLink({
       variables: { importId, target: RegistryDownloadTarget.ImportFile }
     });
-    await downloadFromSignedUrl(link.data?.registryDownloadSignedUrl.signedUrl);
+    downloadFromSignedUrl(link.data?.registryDownloadSignedUrl.signedUrl);
   }
 
   const totalCount = data?.registryImports.totalCount;
